Each user is now saved while they are posting their job hence only the relevent user can update and delete their job post.

+++ backend/job/views.py
@@ -79,6 +79,10 @@ def updateJob(request , pk):
 @api_view(['DELETE']) 
 def deleteJob(request , pk):
     job = get_object_or_404(Job , id = pk)
+
+    if job.user != request.user: 
+        return Response({ 'message' : 'You cannot delete this job'}, status=status.HTTP_403_FORBIDDEN)
+    
     job.delete()
     return Response({'message' : 'Job is deleted Successfully.'}, status=status.HTTP_200_OK)
